refactor(events,notifications): split ticket.py and test_ticket_templates.py under the 1000-line gate
Task 6-9's accumulated additions pushed events/models/ticket.py to 1001
lines and src/notifications/tests/test_ticket_templates.py to 1069,
tripping make check's file-length gate (pre-existing debt, not
introduced by this PR's own diff).

- Move the self-contained Payment model + its Meta/methods out of
  ticket.py into events/models/payment.py (no schema change: Django
  models register by app_label, not file path; verified with
  makemigrations --check). ticket.py re-exports Payment at the bottom
  of the file (after Ticket is defined, to avoid a circular import)
  since ~20 call sites still do `from events.models.ticket import
  Payment`. The FK to Ticket uses the "events.Ticket" string form
  (matching refund.py's existing precedent) instead of importing the
  class, sidestepping any import-order concern.
- Split the Google Wallet email-link tests (Task 8) out of
  test_ticket_templates.py into test_ticket_templates_google_wallet.py.
  The shared ticket_holder/ticket_organization/ticket_event/ticket_tier/
  active_ticket/pending_ticket fixtures and the
  _create_notification_for_test helper move to conftest.py so both
  files can use them without the cross-module fixture-import trick
  (which trips ruff's F811 on every parameter name shadowing the
  import).
