+++ src/notifications/tests/test_ticket_templates.py
@@ -1,20 +1,15 @@
 """Tests for ticket notification templates and attachment generation."""
 
 import base64
-import typing as t
 import uuid
-from datetime import timedelta
-from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
-from django.utils import timezone
 
 from accounts.models import RevelUser
-from events.models import Event, Organization, Ticket, TicketTier
+from events.models import Event, Ticket
 from events.models.ticket import CancellationSource
 from notifications.enums import NotificationType
-from notifications.models import Notification
 from notifications.service.templates.ticket_templates import (
     PaymentConfirmationTemplate,
     TicketCheckedInTemplate,
@@ -27,110 +22,9 @@ from notifications.service.templates.ticket_templates import (
     _load_event,
     _load_ticket,
 )
+from notifications.tests.conftest import _create_notification_for_test
 
 pytestmark = pytest.mark.django_db
-
-
-# --- Fixtures ---
-
-
-@pytest.fixture
-def ticket_holder(django_user_model: type[RevelUser]) -> RevelUser:
-    """A user who holds a ticket."""
-    return django_user_model.objects.create_user(
-        username="holder@example.com",
-        email="holder@example.com",
-        password="password",
-        first_name="Ticket",
-        last_name="Holder",
-    )
-
-
-@pytest.fixture
-def ticket_organization(ticket_holder: RevelUser) -> Organization:
-    """Organization for ticket tests."""
-    return Organization.objects.create(
-        name="Ticket Org",
-        slug="ticket-org",
-        owner=ticket_holder,
-    )
-
-
-@pytest.fixture
-def ticket_event(ticket_organization: Organization) -> Event:
-    """Event for ticket tests."""
-    next_week = timezone.now() + timedelta(days=7)
-    return Event.objects.create(
-        organization=ticket_organization,
-        name="Ticket Event",
-        slug="ticket-event",
-        visibility=Event.Visibility.PUBLIC,
-        event_type=Event.EventType.PUBLIC,
-        max_attendees=100,
-        status="open",
-        start=next_week,
-        end=next_week + timedelta(hours=3),
-        requires_ticket=True,
-    )
-
-
-@pytest.fixture
-def ticket_tier(ticket_event: Event) -> TicketTier:
-    """Ticket tier for tests.
-
-    When an event is created with requires_ticket=True, a default tier is
-    automatically created via signals. We return that tier instead of
-    creating a new one to avoid unique constraint violations.
-    """
-    return ticket_event.ticket_tiers.first()  # type: ignore[return-value]
-
-
-@pytest.fixture
-def active_ticket(
-    ticket_holder: RevelUser,
-    ticket_event: Event,
-    ticket_tier: TicketTier,
-) -> Ticket:
-    """An active ticket for testing."""
-    return Ticket.objects.create(
-        guest_name="Test Guest",
-        user=ticket_holder,
-        event=ticket_event,
-        tier=ticket_tier,
-        status=Ticket.TicketStatus.ACTIVE,
-    )
-
-
-@pytest.fixture
-def pending_ticket(
-    ticket_holder: RevelUser,
-    ticket_event: Event,
-    ticket_tier: TicketTier,
-) -> Ticket:
-    """A pending ticket for testing."""
-    return Ticket.objects.create(
-        guest_name="Test Guest",
-        user=ticket_holder,
-        event=ticket_event,
-        tier=ticket_tier,
-        status=Ticket.TicketStatus.PENDING,
-    )
-
-
-def _create_notification_for_test(
-    user: RevelUser,
-    notification_type: NotificationType,
-    context: dict[str, object],
-) -> Notification:
-    """Create a notification directly without context validation.
-
-    This is for unit testing templates where we only need specific context fields.
-    """
-    return Notification.objects.create(
-        user=user,
-        notification_type=notification_type,
-        context=context,
-    )
 
 
 # --- _load_ticket Tests ---
@@ -940,130 +834,3 @@ class TestPendingTicketPaymentInstructions:
         rendered = self._render(template, instructions=None)
 
         assert "contact the organizer" in rendered
-
-
-# --- Google Wallet Save Link Tests ---
-
-
-@pytest.fixture
-def google_wallet_settings(settings: t.Any, tmp_path: Path) -> None:
-    """Configure Google Wallet with a real (test-only) service-account key."""
-    import json
-
-    from cryptography.hazmat.primitives import serialization
-    from cryptography.hazmat.primitives.asymmetric import rsa
-
-    key = rsa.generate_private_key(public_exponent=65537, key_size=2048)
-    pem = key.private_bytes(
-        encoding=serialization.Encoding.PEM,
-        format=serialization.PrivateFormat.PKCS8,
-        encryption_algorithm=serialization.NoEncryption(),
-    ).decode()
-    sa_path = tmp_path / "sa.json"
-    sa_path.write_text(json.dumps({"client_email": "wallet@test.iam.gserviceaccount.com", "private_key": pem}))
-    settings.GOOGLE_WALLET_ISSUER_ID = "3388000000012345678"
-    settings.GOOGLE_WALLET_SERVICE_ACCOUNT_KEY_PATH = str(sa_path)
-    settings.GOOGLE_WALLET_CLASS_PREFIX = "test"
-
-
-class TestGoogleWalletEmailLink:
-    """Tests for the Google Wallet save link in ticket emails."""
-
-    @staticmethod
-    def _notification(ticket_holder: RevelUser, active_ticket: Ticket, **extra: object) -> Notification:
-        return _create_notification_for_test(
-            user=ticket_holder,
-            notification_type=NotificationType.TICKET_CREATED,
-            context={
-                "event_name": active_ticket.event.name,
-                "ticket_status": "active",
-                "ticket_id": str(active_ticket.id),
-                "event_id": str(active_ticket.event.id),
-                **extra,
-            },
-        )
-
-    def test_link_present_when_configured(
-        self, ticket_holder: RevelUser, active_ticket: Ticket, google_wallet_settings: None
-    ) -> None:
-        notification = self._notification(ticket_holder, active_ticket)
-        template = TicketCreatedTemplate()
-
-        html = template.get_email_html_body(notification)
-        text = template.get_email_text_body(notification)
-
-        assert html is not None and "https://pay.google.com/gp/v/save/" in html
-        assert "https://pay.google.com/gp/v/save/" in text
-
-    def test_link_absent_when_unconfigured(
-        self, ticket_holder: RevelUser, active_ticket: Ticket, settings: t.Any
-    ) -> None:
-        settings.GOOGLE_WALLET_ISSUER_ID = ""
-        settings.GOOGLE_WALLET_SERVICE_ACCOUNT_KEY_PATH = ""
-        notification = self._notification(ticket_holder, active_ticket)
-        template = TicketCreatedTemplate()
-
-        html = template.get_email_html_body(notification)
-
-        assert html is not None and "pay.google.com" not in html
-
-    def test_link_absent_when_include_pkpass_false(
-        self, ticket_holder: RevelUser, active_ticket: Ticket, google_wallet_settings: None
-    ) -> None:
-        notification = self._notification(ticket_holder, active_ticket, include_pkpass=False)
-        template = TicketCreatedTemplate()
-
-        html = template.get_email_html_body(notification)
-
-        assert html is not None and "pay.google.com" not in html
-
-    def test_link_absent_when_ticket_cancelled(
-        self, ticket_holder: RevelUser, active_ticket: Ticket, google_wallet_settings: None
-    ) -> None:
-        """TICKET_CANCELLED/TICKET_REFUNDED reuse TicketUpdatedTemplate with include_pkpass
-
-        defaulting True — a cancelled ticket must never get a save link, mirroring
-        TicketWalletController.get_queryset()'s ACTIVE/PENDING status filter.
-        """
-        active_ticket.status = Ticket.TicketStatus.CANCELLED
-        active_ticket.save(update_fields=["status"])
-        notification = _create_notification_for_test(
-            user=ticket_holder,
-            notification_type=NotificationType.TICKET_UPDATED,
-            context={
-                "event_name": active_ticket.event.name,
-                "action": "cancelled",
-                "ticket_status": "cancelled",
-                "ticket_id": str(active_ticket.id),
-                "event_id": str(active_ticket.event.id),
-            },
-        )
-        template = TicketUpdatedTemplate()
-
-        html = template.get_email_html_body(notification)
-
-        assert html is not None and "pay.google.com" not in html
-
-    def test_ticket_updated_activation_link_present_when_configured(
-        self, ticket_holder: RevelUser, active_ticket: Ticket, google_wallet_settings: None
-    ) -> None:
-        """Pins the TicketUpdatedTemplate placement in the pending->active activation branch."""
-        notification = _create_notification_for_test(
-            user=ticket_holder,
-            notification_type=NotificationType.TICKET_UPDATED,
-            context={
-                "event_name": active_ticket.event.name,
-                "old_status": "pending",
-                "new_status": "active",
-                "ticket_status": "active",
-                "ticket_id": str(active_ticket.id),
-                "event_id": str(active_ticket.event.id),
-            },
-        )
-        template = TicketUpdatedTemplate()
-
-        html = template.get_email_html_body(notification)
-        text = template.get_email_text_body(notification)
-
-        assert html is not None and "https://pay.google.com/gp/v/save/" in html
-        assert "https://pay.google.com/gp/v/save/" in text
